Renamed index to page so we don't have 10 indextests
