+++ src/LinkDotNet.Blog.Web/ServiceExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using LinkDotNet.Blog.Web.Features.Admin.BlogPostEditor.Services;
 using LinkDotNet.Blog.Web.Features.Admin.Dashboard.Services;
-using LinkDotNet.Blog.Web.Features.Admin.Sitemap;
 using LinkDotNet.Blog.Web.Features.Admin.Sitemap.Services;
 using LinkDotNet.Blog.Web.Features.Services;
 using Microsoft.Extensions.DependencyInjection;
