+++ tests/LinkDotNet.Blog.IntegrationTests/Web/Pages/Admin/DraftBlogPostPageTests.cs
@@ -22,7 +22,7 @@ public class DraftBlogPostPageTests : SqlDatabaseTestBase<BlogPost>
         using var ctx = new TestContext();
         ctx.JSInterop.Mode = JSRuntimeMode.Loose;
         ctx.Services.AddScoped<IRepository<BlogPost>>(_ => Repository);
-        var cut = ctx.RenderComponent<Index>();
+        var cut = ctx.RenderComponent<DraftBlogPostPage>();
         cut.WaitForState(() => cut.FindAll(".blog-card").Any());
 
         var blogPosts = cut.FindComponents<ShortBlogPost>();
