+++ tests/LinkDotNet.Blog.UnitTests/Web/Shared/Services/XmlFileWriterTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Threading.Tasks;
-using LinkDotNet.Blog.Web.Features.Admin.Sitemap;
+using LinkDotNet.Blog.Web.Features.Admin.Sitemap.Services;
 
 namespace LinkDotNet.Blog.UnitTests.Web.Shared.Services;
